moved PermissionChecker into util module
